fixes to mypy and tests
