+++ src/napari/components/_tests/test_viewer_mouse_bindings.py
@@ -1,4 +1,3 @@
-import os
 from unittest.mock import Mock
 
 import numpy as np
@@ -138,59 +137,3 @@ def test_double_click_to_zoom(layer_shape):
     # Assert nothing has changed
     assert viewer.camera.zoom == initial_zoom
     assert np.allclose(viewer.camera.center, (0, 0, 0))
-
-
-def test_viewer_drag_to_zoom(qtbot, make_napari_viewer):
-    """Test drag to zoom mouse binding."""
-    np.random.seed(0)
-    viewer = make_napari_viewer()
-    canvas = viewer.window._qt_viewer.canvas
-
-    if os.getenv('CI'):
-        viewer.show()
-
-    def zoom_callback(event):
-        """Mock zoom callback to check zoom box visibility."""
-        data_positions = event.value
-        assert len(data_positions) == 2, (
-            'Zoom event should release two positions'
-        )
-
-    viewer._zoom_box.events.zoom.connect(zoom_callback)
-
-    # Add an image layer
-    data = np.random.random((10, 20))
-    viewer.add_image(data)
-
-    assert viewer._zoom_box.visible is False, (
-        'Zoom box should be hidden initially'
-    )
-    # Simulate press to start zooming
-    canvas._scene_canvas.events.mouse_press(
-        pos=(0, 0), modifiers=('Alt',), button=0
-    )
-    qtbot.wait(10)
-    assert viewer._zoom_box.visible is True, (
-        'Zoom box should be visible after press'
-    )
-
-    # Simulate drag to zoom
-    canvas._scene_canvas.events.mouse_move(
-        pos=(100, 100), modifiers=('Alt',), button=0, press_event=True
-    )
-    qtbot.wait(10)
-    assert viewer._zoom_box.visible is True, (
-        'Zoom box should remain visible during drag'
-    )
-    assert viewer._zoom_box.canvas_positions == ((0, 0), (100, 100)), (
-        'Zoom box canvas positions should match the drag coordinates'
-    )
-
-    # Simulate release to finish zooming
-    canvas._scene_canvas.events.mouse_release(
-        pos=(100, 100), modifiers=('Alt',), button=0
-    )
-    qtbot.wait(10)
-    assert viewer._zoom_box.visible is False, (
-        'Zoom box should be hidden after release'
-    )
